Move Storybook stories to root stories directory
